backend/requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
tensorflow-cpu
joblib
